Use MUnit in ActorSuite
